requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
PyYAML==6.0
Werkzeug==2.3.8
